[refactor] auth: 불필요한 서비스 인터페이스 삭제 및 통합
