+++ src/main/java/org/example/ootoutfitoftoday/domain/auth/service/command/AuthCommandService.java
@@ -1,17 +1,81 @@
 package org.example.ootoutfitoftoday.domain.auth.service.command;
 
+import jakarta.transaction.Transactional;
+import lombok.RequiredArgsConstructor;
 import org.example.ootoutfitoftoday.domain.auth.dto.request.AuthSignupRequest;
+import org.example.ootoutfitoftoday.domain.auth.exception.AuthErrorCode;
+import org.example.ootoutfitoftoday.domain.auth.exception.AuthException;
+import org.example.ootoutfitoftoday.domain.user.entity.User;
+import org.example.ootoutfitoftoday.domain.user.enums.UserRole;
+import org.example.ootoutfitoftoday.domain.user.service.command.UserCommandService;
+import org.example.ootoutfitoftoday.domain.user.service.query.UserQueryService;
+import org.example.ootoutfitoftoday.security.jwt.JwtUtil;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
 
-public interface AuthCommandService {
+@Service
+@RequiredArgsConstructor
+@Transactional
+public class AuthCommandService {
 
-    void initializeAdmin(
+    private final UserCommandService userCommandService;
+    private final UserQueryService userQueryService;
+    private final JwtUtil jwtUtil;
+    private final PasswordEncoder passwordEncoder;
+
+    // 관리자 계정 초기 생성 자동
+    public void initializeAdmin(
             String loginId,
             String email,
             String nickname,
             String username,
             String password,
             String phoneNumber
-    );
+    ) {
+        if (!userQueryService.existsByLoginId(loginId)) {
+            User admin = User.createAdmin(
+                    loginId,
+                    email,
+                    nickname,
+                    username,
+                    passwordEncoder.encode(password),
+                    phoneNumber
+            );
+            userCommandService.save(admin);
+            System.out.println("관리자 계정 초기 생성 완료되었습니다.");
+        } else {
+            System.out.println("관리자 계정이 이미 존재합니다.");
+        }
+    }
 
-    void signup(AuthSignupRequest request);
+    // 회원가입
+    public void signup(AuthSignupRequest request) {
+
+        if (userQueryService.existsByLoginId(request.getLoginId())) {
+            throw new AuthException(AuthErrorCode.DUPLICATE_LOGIN_ID);
+        }
+        if (userQueryService.existsByEmail(request.getEmail())) {
+            throw new AuthException(AuthErrorCode.DUPLICATE_EMAIL);
+        }
+        if (userQueryService.existsByNickname(request.getNickname())) {
+            throw new AuthException(AuthErrorCode.DUPLICATE_NICKNAME);
+        }
+        if (userQueryService.existsByPhoneNumber(request.getPhoneNumber())) {
+            throw new AuthException(AuthErrorCode.DUPLICATE_PHONE_NUMBER);
+        }
+
+        String encodedPassword = passwordEncoder.encode(request.getPassword());
+
+        User user = User.builder()
+                .loginId(request.getLoginId())
+                .email(request.getEmail())
+                .nickname(request.getNickname())
+                .username(request.getUsername())
+                .password(encodedPassword)
+                .phoneNumber(request.getPhoneNumber())
+                .role(UserRole.ROLE_USER)
+                .build();
+
+        userCommandService.save(user);
+    }
 }
